+++ KinaUnaProgenyApi/Controllers/ContactsController.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using KinaUna.Data;
 using KinaUna.Data.Extensions;
 using KinaUna.Data.Models;
+using KinaUnaProgenyApi.Services.UserAccessService;
 
 namespace KinaUnaProgenyApi.Controllers
 {
